feat: 新增創建訂單時用折扣碼去計算額外的折扣、調整金額 amount 的型態為BigDecimal

+++ springboot-mall/src/main/java/com/chunyue/springbootmall/dao/OrderDao.java
@@ -4,11 +4,12 @@ import com.chunyue.springbootmall.dto.QueryParams;
 import com.chunyue.springbootmall.model.Order;
 import com.chunyue.springbootmall.model.OrderItem;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface OrderDao {
 
-    Integer createOrder(Integer userId, Integer amount);
+    Integer createOrder(Integer userId, BigDecimal amount);
 
     void createOrderItems(Integer orderId, List<OrderItem> list);
